% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_year.R
\name{corr_year}
\alias{corr_year}
\title{Correlogram Function}
\usage{
corr_year(year)
}
\arguments{
\item{year}{What year?}
}
\description{
This function allows you to produce the correlogram for a certain NBA year.
}
\examples{
corr_year(1996)
}
\keyword{1950}
\keyword{2013}
\keyword{from}
\keyword{to}
