% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_team.R
\name{find_team}
\alias{find_team}
\title{Find Team Function}
\usage{
find_team(year)
}
\arguments{
\item{year}{What year do you want to look at?}
}
\description{
This function allows you to find the team NBA members were part of for a specific year
}
\examples{
find_team(1995)
}
\keyword{1950}
\keyword{2013}
\keyword{from}
\keyword{to}
