% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{Find Oldest Player Function}
\usage{
oldest_player(year)
}
\arguments{
\item{year}{What year?}
}
\description{
This function allows you to find the oldest NBA player of that year
}
\examples{
oldest_player(1996)
}
\keyword{1950}
\keyword{2013}
\keyword{from}
\keyword{to}
